fixed setegid(), was non functional before; fixed more warnings with Sun Studio compilers
git-svn-id: svn+ssh://svn.code.sf.net/p/qore/code/qore/trunk@3722 1b0b35fe-f740-4c1b-b404-a338b32e513b

+++ lib/QC_FtpClient.cpp
@@ -154,7 +154,7 @@ static AbstractQoreNode *FC_setHostName(QoreObject *self, QoreFtpClientClass *f,
 }
 
 static AbstractQoreNode *FC_setPort(QoreObject *self, QoreFtpClientClass *f, const QoreListNode *args, ExceptionSink *xsink) {
-   int port = HARD_QORE_INT(args, 0);
+   int port = (int)HARD_QORE_INT(args, 0);
    if (port <= 0) {
       xsink->raiseException("FTPCLIENT-SETPORT-PARAMETER-ERROR", "expecting positive port number as first parameter of FtpClient::setPort(softint $port); got %d", port);
       return 0;
